# More efficient SortedArray.insert(contentsOf:)