--- a/Sources/Basic/SortedArray.swift
+++ b/Sources/Basic/SortedArray.swift
@@ -59,14 +59,55 @@ public struct SortedArray<Element>: CustomStringConvertible {
     
     /// Insert the given sequence, maintaining the sort order.
     public mutating func insert<S: Sequence>(contentsOf newElements: S) where S.Iterator.Element == Element {
+        var newElements = newElements.sorted(by: areInIncreasingOrder)
+        guard !newElements.isEmpty else {
+            return
+        }
+        guard !elements.isEmpty else {
+            elements = newElements
+            return
+        }
+
+        var lhsIndex = elements.endIndex - 1
+        var rhsIndex = newElements.endIndex - 1
+
+        elements.reserveCapacity(elements.count + newElements.count)
+
+        // TODO: If SortedArray moves to stdlib an _ArrayBuffer can be used
+        // instead, this can then be removed as _ArrayBuffer can be resized
+        // without requiring instantiated elements.
         elements.append(contentsOf: newElements)
-        elements.sort(by: areInIncreasingOrder)
+
+        var lhs = elements[lhsIndex], rhs = newElements[rhsIndex]
+
+        // Equivalent to a merge sort, "pop" and append the max elemeent of
+        // each array until either array is empty
+        for index in elements.indices.reversed() {
+            if areInIncreasingOrder(lhs, rhs) {
+                elements[index] = rhs
+                rhsIndex -= 1
+                guard rhsIndex >= 0 else { break }
+                rhs = newElements[rhsIndex]
+            }
+            else {
+                elements[index] = lhs
+                lhsIndex -= 1
+                guard lhsIndex >= 0 else { break }
+                lhs = elements[lhsIndex]
+            }
+        }
+        // any remaining new elements were smaller than all old elements
+        // so the remaining new elements can safely replace the prefix.
+        if rhsIndex >= 0 {
+            elements.replaceSubrange(0 ... rhsIndex,
+                                     with: newElements[0 ... rhsIndex])
+        }
     }
 
-	/// Returns the values as an array.
-	public var values: [Element] {
-		return elements
-	}
+    /// Returns the values as an array.
+    public var values: [Element] {
+        return elements
+    }
     
     public var description: String {
         return "<SortedArray: \(elements)>"