--- a/Tests/BasicPerformanceTests/SortedArrayPerfTests.swift
+++ b/Tests/BasicPerformanceTests/SortedArrayPerfTests.swift
@@ -11,7 +11,17 @@
 import TestSupport
 import Basic
 
+private let magnitude = 5000
+
+private func prng(_ seed: Int) -> Int {
+    return (seed &* 1299827) % magnitude
+}
+
 class SortedArrayPerfTests: XCTestCasePerf {
+
+    private let firstSequence = (0..<(magnitude*2)).map(prng)
+    private let secondSequence = ((magnitude*1)..<(magnitude*3)).map(prng)
+
     func testPerformanceOfSortedArrayInAscendingOrder() {
         
         measure() {
@@ -19,6 +29,32 @@ class SortedArrayPerfTests: XCTestCasePerf {
             for i in 1...1000 {
                 arr.insert(i)
             }
+        }
+    }
+
+    func testPerformanceOfSortedArrayInsertWithDuplicates() {
+        @inline(never)
+        func insert<T, S: Sequence>(_ sequence: S, into array: inout SortedArray<T>) where S.Iterator.Element == T {
+            for element in sequence {
+                array.insert(element)
+            }
+        }
+
+        var arr = SortedArray<Int>(areInIncreasingOrder: <)
+        arr.insert(contentsOf: firstSequence)
+
+        measure() {
+            insert(self.secondSequence, into: &arr)
+        }
+    }
+
+    func testPerformanceOfSortedArrayInsertContentsOfWithDuplicates() {
+
+        var arr = SortedArray<Int>(areInIncreasingOrder: <)
+        arr.insert(contentsOf: firstSequence)
+
+        measure() {
+            arr.insert(contentsOf: self.secondSequence)
         }
     }
 }