Uses the equivalent of a single pass of a merge-sort to make the `SortedArray.insert(contentsOf:)` more efficient.

This allows the algorithm to go from `O((a+b) log(a+b))` to `O(a + b log (b))`. Where `a` is the size of the current array, and `b` is the size of the new elements. If `b` is already sorted (i.e. another `SortedArray`) then an efficient `sort` function will mean this runs in `O(a + b)`.

The new performance tests should demonstrate the improvements when compared, assuming the element count is suitably large:
> testPerformanceOfSortedArrayInsertWithDuplicates()
> testPerformanceOfSortedArrayInsertContentsOfWithDuplicates()

**Note** I needed to add `-swift-version 3` to `Other Swift Flags`

**Note** Insertion times are negligible for a `magnitude` of less than 1000. Performance on my machine is actually slightly worse for a `magnitude` less than 1000. However it gets significantly better for values of `magnitude` higher than 1000. I suspect this is because of `elements.append(contentsOf: newElements)` and constant overheads.