@aciidb0mb3r 2017-05-18
Can you also post the perf results with and without your patch?