@therealbnut 2017-05-18
**Before Patch**
```
testPerformanceOfSortedArrayInsertWithDuplicates: 0.030 seconds
testPerformanceOfSortedArrayInsertContentsOfWithDuplicates: 0.033 seconds
```

**After Patch**
```
testPerformanceOfSortedArrayInsertWithDuplicates: 0.030 seconds
testPerformanceOfSortedArrayInsertContentsOfWithDuplicates: 0.026 seconds
```

So `0.788%` of the original time.

**Raw output Before**
```
Test Suite 'BasicPerformanceTests.xctest' started at 2017-05-19 07:36:26.510
Test Suite 'SortedArrayPerfTests' started at 2017-05-19 07:36:26.510
Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInAscendingOrder]' started.
<unknown>:0: Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInAscendingOrder]' measured [Time, seconds] average: 0.001, relative standard deviation: 22.328%, values: [0.001628, 0.001218, 0.000962, 0.000961, 0.000961, 0.000939, 0.000854, 0.000893, 0.000884, 0.000857], performanceMetricID:com.apple.XCTPerformanceMetric_WallClockTime, baselineName: "", baselineAverage: , maxPercentRegression: 10.000%, maxPercentRelativeStandardDeviation: 10.000%, maxRegression: 0.100, maxStandardDeviation: 0.100
Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInAscendingOrder]' passed (0.400 seconds).
Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInsertContentsOfWithDuplicates]' started.
<unknown>:0: Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInsertContentsOfWithDuplicates]' measured [Time, seconds] average: 0.033, relative standard deviation: 13.649%, values: [0.039102, 0.029542, 0.026896, 0.028125, 0.031120, 0.037950, 0.040461, 0.034373, 0.035334, 0.030393], performanceMetricID:com.apple.XCTPerformanceMetric_WallClockTime, baselineName: "", baselineAverage: , maxPercentRegression: 10.000%, maxPercentRelativeStandardDeviation: 10.000%, maxRegression: 0.100, maxStandardDeviation: 0.100
Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInsertContentsOfWithDuplicates]' passed (0.590 seconds).
Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInsertWithDuplicates]' started.
<unknown>:0: Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInsertWithDuplicates]' measured [Time, seconds] average: 0.030, relative standard deviation: 10.156%, values: [0.036232, 0.029010, 0.029690, 0.029673, 0.033373, 0.025381, 0.030920, 0.026916, 0.028370, 0.027521], performanceMetricID:com.apple.XCTPerformanceMetric_WallClockTime, baselineName: "", baselineAverage: , maxPercentRegression: 10.000%, maxPercentRelativeStandardDeviation: 10.000%, maxRegression: 0.100, maxStandardDeviation: 0.100
Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInsertWithDuplicates]' passed (0.550 seconds).
Test Suite 'SortedArrayPerfTests' passed at 2017-05-19 07:36:28.052.
	 Executed 3 tests, with 0 failures (0 unexpected) in 1.540 (1.541) seconds
Test Suite 'BasicPerformanceTests.xctest' passed at 2017-05-19 07:36:28.052.
	 Executed 3 tests, with 0 failures (0 unexpected) in 1.540 (1.542) seconds
```

**Raw output After**
```
Test Suite 'BasicPerformanceTests.xctest' started at 2017-05-19 07:28:24.253
Test Suite 'SortedArrayPerfTests' started at 2017-05-19 07:28:24.253
Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInAscendingOrder]' started.
<unknown>:0: Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInAscendingOrder]' measured [Time, seconds] average: 0.001, relative standard deviation: 22.369%, values: [0.001597, 0.000966, 0.000927, 0.000856, 0.000856, 0.000856, 0.000905, 0.000860, 0.000902, 0.000881], performanceMetricID:com.apple.XCTPerformanceMetric_WallClockTime, baselineName: "", baselineAverage: , maxPercentRegression: 10.000%, maxPercentRelativeStandardDeviation: 10.000%, maxRegression: 0.100, maxStandardDeviation: 0.100
Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInAscendingOrder]' passed (0.409 seconds).
Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInsertContentsOfWithDuplicates]' started.
<unknown>:0: Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInsertContentsOfWithDuplicates]' measured [Time, seconds] average: 0.026, relative standard deviation: 5.713%, values: [0.027127, 0.026573, 0.024120, 0.025614, 0.023618, 0.027691, 0.023619, 0.026337, 0.026330, 0.024002], performanceMetricID:com.apple.XCTPerformanceMetric_WallClockTime, baselineName: "", baselineAverage: , maxPercentRegression: 10.000%, maxPercentRelativeStandardDeviation: 10.000%, maxRegression: 0.100, maxStandardDeviation: 0.100
Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInsertContentsOfWithDuplicates]' passed (0.509 seconds).
Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInsertWithDuplicates]' started.
<unknown>:0: Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInsertWithDuplicates]' measured [Time, seconds] average: 0.030, relative standard deviation: 6.425%, values: [0.031083, 0.028376, 0.028500, 0.030827, 0.028190, 0.031095, 0.030869, 0.032336, 0.027447, 0.033804], performanceMetricID:com.apple.XCTPerformanceMetric_WallClockTime, baselineName: "", baselineAverage: , maxPercentRegression: 10.000%, maxPercentRelativeStandardDeviation: 10.000%, maxRegression: 0.100, maxStandardDeviation: 0.100
Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInsertWithDuplicates]' passed (0.554 seconds).
Test Suite 'SortedArrayPerfTests' passed at 2017-05-19 07:28:25.727.
	 Executed 3 tests, with 0 failures (0 unexpected) in 1.472 (1.473) seconds
Test Suite 'BasicPerformanceTests.xctest' passed at 2017-05-19 07:28:25.727.
	 Executed 3 tests, with 0 failures (0 unexpected) in 1.472 (1.474) seconds
```