@therealbnut 2017-05-18
Worth noting, if `magnitude` is changed to `500`, then the output is:

**Before Patch**
```
testPerformanceOfSortedArrayInsertWithDuplicates: 0.002 seconds
testPerformanceOfSortedArrayInsertContentsOfWithDuplicates: 0.003 seconds
```

**After Patch**
```
testPerformanceOfSortedArrayInsertWithDuplicates: 0.002 seconds
testPerformanceOfSortedArrayInsertContentsOfWithDuplicates: 0.004 seconds
```

If however we fall back to the old approach when `elements` is less than 1000 then this works fine (see 2266a40):
```
testPerformanceOfSortedArrayInsertWithDuplicates: 0.002 seconds
testPerformanceOfSortedArrayInsertContentsOfWithDuplicates: 0.002 seconds
```

However the derivation of `1000` is a little dodgy.

Note that this performs a little better than it used to, probably because:
```diff
- elements.append(contentsOf: newElements)
+ elements.append(contentsOf: newElements.sorted(by: areInIncreasingOrder))
  elements.sort(by: areInIncreasingOrder)
```