@therealbnut 2017-05-18
By moving the initialisation and starting data into the measure block we are also measuring the complexity of sorting that original data. So it goes from being `O(a+b log(b))` to `O(a log(a) + a + b log(b))`, this may make it hard to see improvements if it's tweaked further.