@therealbnut 2017-05-18
**Before Patch**
```
testPerformanceOfSortedArrayInAscendingOrder: 0.003 seconds
testPerformanceOfSortedArrayInsertWithDuplicates: 0.063 seconds
testPerformanceOfSortedArrayInsertContentsOfWithDuplicates: 0.033 seconds
```

**After Patch**
```
testPerformanceOfSortedArrayInAscendingOrder: 0.004 seconds
testPerformanceOfSortedArrayInsertWithDuplicates: 0.067 seconds
testPerformanceOfSortedArrayInsertContentsOfWithDuplicates: 0.026 seconds
```

<details>
<summary><b>Raw output</b></summary>

**Before**

```
Test Suite 'BasicPerformanceTests.xctest' started at 2017-05-19 09:42:53.671
Test Suite 'SortedArrayPerfTests' started at 2017-05-19 09:42:53.672
Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInAscendingOrder]' started.
<unknown>:0: Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInAscendingOrder]' measured [Time, seconds] average: 0.003, relative standard deviation: 20.383%, values: [0.003794, 0.002829, 0.002442, 0.002333, 0.002206, 0.002250, 0.002311, 0.003675, 0.003311, 0.002992], performanceMetricID:com.apple.XCTPerformanceMetric_WallClockTime, baselineName: "", baselineAverage: , maxPercentRegression: 10.000%, maxPercentRelativeStandardDeviation: 10.000%, maxRegression: 0.100, maxStandardDeviation: 0.100
Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInAscendingOrder]' passed (0.431 seconds).
Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInsertContentsOfWithDuplicates]' started.
<unknown>:0: Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInsertContentsOfWithDuplicates]' measured [Time, seconds] average: 0.060, relative standard deviation: 43.386%, values: [0.022566, 0.030910, 0.038230, 0.046657, 0.053338, 0.057295, 0.069899, 0.078052, 0.088518, 0.109850], performanceMetricID:com.apple.XCTPerformanceMetric_WallClockTime, baselineName: "", baselineAverage: , maxPercentRegression: 10.000%, maxPercentRelativeStandardDeviation: 10.000%, maxRegression: 0.100, maxStandardDeviation: 0.100
Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInsertContentsOfWithDuplicates]' passed (0.863 seconds).
Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInsertWithDuplicates]' started.
<unknown>:0: Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInsertWithDuplicates]' measured [Time, seconds] average: 0.063, relative standard deviation: 52.654%, values: [0.017970, 0.025242, 0.031361, 0.042096, 0.054360, 0.064290, 0.076851, 0.091277, 0.108693, 0.117150], performanceMetricID:com.apple.XCTPerformanceMetric_WallClockTime, baselineName: "", baselineAverage: , maxPercentRegression: 10.000%, maxPercentRelativeStandardDeviation: 10.000%, maxRegression: 0.100, maxStandardDeviation: 0.100
Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInsertWithDuplicates]' passed (0.896 seconds).
Test Suite 'SortedArrayPerfTests' passed at 2017-05-19 09:42:55.862.
	 Executed 3 tests, with 0 failures (0 unexpected) in 2.190 (2.191) seconds
Test Suite 'BasicPerformanceTests.xctest' passed at 2017-05-19 09:42:55.863.
	 Executed 3 tests, with 0 failures (0 unexpected) in 2.190 (2.191) seconds
```

**Raw output After**
```
Test Suite 'BasicPerformanceTests.xctest' started at 2017-05-19 09:45:47.595
Test Suite 'SortedArrayPerfTests' started at 2017-05-19 09:45:47.596
Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInAscendingOrder]' started.
<unknown>:0: Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInAscendingOrder]' measured [Time, seconds] average: 0.004, relative standard deviation: 10.672%, values: [0.002943, 0.003683, 0.003919, 0.003859, 0.003769, 0.003439, 0.003296, 0.003545, 0.004154, 0.004366], performanceMetricID:com.apple.XCTPerformanceMetric_WallClockTime, baselineName: "", baselineAverage: , maxPercentRegression: 10.000%, maxPercentRelativeStandardDeviation: 10.000%, maxRegression: 0.100, maxStandardDeviation: 0.100
Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInAscendingOrder]' passed (0.431 seconds).
Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInsertContentsOfWithDuplicates]' started.
<unknown>:0: Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInsertContentsOfWithDuplicates]' measured [Time, seconds] average: 0.026, relative standard deviation: 20.954%, values: [0.018356, 0.021547, 0.022442, 0.023262, 0.024341, 0.023504, 0.026759, 0.027805, 0.034292, 0.036708], performanceMetricID:com.apple.XCTPerformanceMetric_WallClockTime, baselineName: "", baselineAverage: , maxPercentRegression: 10.000%, maxPercentRelativeStandardDeviation: 10.000%, maxRegression: 0.100, maxStandardDeviation: 0.100
Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInsertContentsOfWithDuplicates]' passed (0.525 seconds).
Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInsertWithDuplicates]' started.
<unknown>:0: Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInsertWithDuplicates]' measured [Time, seconds] average: 0.067, relative standard deviation: 49.136%, values: [0.023819, 0.028523, 0.033497, 0.044978, 0.058423, 0.071664, 0.079276, 0.098499, 0.110367, 0.118998], performanceMetricID:com.apple.XCTPerformanceMetric_WallClockTime, baselineName: "", baselineAverage: , maxPercentRegression: 10.000%, maxPercentRelativeStandardDeviation: 10.000%, maxRegression: 0.100, maxStandardDeviation: 0.100
Test Case '-[BasicPerformanceTests.SortedArrayPerfTests testPerformanceOfSortedArrayInsertWithDuplicates]' passed (0.934 seconds).
Test Suite 'SortedArrayPerfTests' passed at 2017-05-19 09:45:49.486.
	 Executed 3 tests, with 0 failures (0 unexpected) in 1.890 (1.891) seconds
Test Suite 'BasicPerformanceTests.xctest' passed at 2017-05-19 09:45:49.487.
	 Executed 3 tests, with 0 failures (0 unexpected) in 1.890 (1.891) seconds
```
</details>